Comments and some moving
Put comments on my code other than buttons. Also put the Controller into its own package, got rid of reduntant files, and did some slight renaming.
