--- ProjectGUI/src/Model/SpellList.java
@@ -1,37 +1,0 @@
-package Model;
-
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-
-/**
- *
- * @author Robert J.
- * 
- * Class: Spell List
- * public
- * 
- * Collaborators:
- *  SpellBook, MenuManager
- * 
- * Responsibilities:
- *  Process and edit a list of valid spells and abilities from a JSON file
- * 
- * Methods:
- *  Spell Enum:
- *      private
- *      holds all available spells
- *  SpellEdit:
- *      private
- *      Edits which spells are available for a character
- *  FileRead:
- *      private
- *      Reads from the JSON file
- */
-public class SpellList {
-    
-
-    
-    
-}
