+++ ProjectGUI/src/View/EditSkills.form
@@ -182,7 +182,7 @@
                               <Component id="jTextField11" min="-2" max="-2" attributes="0"/>
                               <Component id="jTextField12" min="-2" max="-2" attributes="0"/>
                           </Group>
-                          <EmptySpace pref="32" max="32767" attributes="0"/>
+                          <EmptySpace pref="34" max="32767" attributes="0"/>
                       </Group>
                   </Group>
               </Group>
